Continued mutex paranoia, added support for symlinked desktop entries
There was a few changes and additions in this one. We now use a two-way mapping system to deal with symlinks and their nuance. Filesystem watching was reworked a little to account for some of the design implications that comes from support symlinks as well. loadFiles() saw some abstraction, namely the introduction of the setVars() function that handles the oneshot var setup, and the abstraction of walkFunction() which was moved out of the load scope to improve readability. Some functions, such as the specific event handlers, now have rudimentary logic documentation (sorry).